requirements.txt
python-dotenv
fastAPI
uvicorn
ifcopenshell
ifctester
